Refactored fetching user from local storage
Moved back to app
