+++ client/src/components/App.tsx
@@ -5,13 +5,31 @@ import ChatPage from './ChatPage';
 
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import { useAppSelector } from '../hooks/typedReduxHooks';
-import { useStoredUser } from '../hooks/useStoredUser';
+import { useAppDispatch, useAppSelector } from '../hooks/typedReduxHooks';
+import { localStorageKeys } from '../utils/constants';
+import { useEffect } from 'react';
+import { userWithTokenSchema } from '../validators/UserWithToken';
+import { setCurrentUser } from '../reducers/userReducer';
 
 const App = () => {
-  useStoredUser();
-
   const currentUser = useAppSelector((state) => state.user.currentUser);
+
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    const localStorageUser = window.localStorage.getItem(localStorageKeys.currentUser);
+    if (!localStorageUser) return;
+
+    const user = userWithTokenSchema.safeParse(JSON.parse(localStorageUser));
+
+    if (!user.success) {
+      console.error('Parsed user invalid');
+      return;
+    }
+
+    dispatch(setCurrentUser(user.data));
+  }, [dispatch]);
+
 
   return (
     <>
